Tag most methods with #[inline]
The ring buffer we provide is ultimately just a thin wrapper around a
slice of data, with most methods being arguably very short (and trivial
in terms of what they do).
Hence, it most definitely makes sense to inline these at the actual call
sites and with this change we do that. However, we just hint at this
possibility, but do not enforce it (as #[inline(always)] would).